+++ CHANGES.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+* Finished tasks are no longer marked as having an uncertain finish date (fixes [#2](https://github.com/derkork/project.txt-js/issues/2)).
+
 ## [0.1.3] - 2020-03-28
 ### Fixed
 * The library now properly calculates whether a finish date has uncertainty when a task has more than one prerequisite.